+++ src/HCFS/macro.h
@@ -96,7 +96,7 @@
 #define FTELL(A)\
 	{\
 		errcode = 0;\
-		ret_pos = ftell(A);\
+		ret_pos = (int64_t) ftell(A);\
 		if (ret_pos < 0) {\
 			errcode = errno;\
 			write_log(0, "IO error in %s. Code %d, %s\n", __func__,\
